Removing a dependency. Platform does not depend on routing anymore.
