+++ map/map_tests/gps_track_collection_test.cpp
@@ -2,6 +2,8 @@
 
 #include "map/gps_track_collection.hpp"
 
+#include "geometry/latlon.hpp"
+
 #include "base/logging.hpp"
 
 #include <chrono>
